+++ src/commands/database-instances-list.ts
@@ -9,7 +9,7 @@ import { needProjectNumber } from "../projectUtils";
 import firedata = require("../gcp/firedata");
 import { Emulators } from "../emulator/types";
 import { warnEmulatorNotSupported } from "../emulator/commandUtils";
-import { previews } from "../previews";
+import * as experiments from "../experiments";
 import { needProjectId } from "../projectUtils";
 import {
   listDatabaseInstances,
@@ -52,7 +52,7 @@ export let command = new Command("database:instances:list")
     ).start();
     let instances;
 
-    if (previews.rtdbmanagement) {
+    if (experiments.isEnabled("rtdbmanagement")) {
       const projectId = needProjectId(options);
       try {
         instances = await listDatabaseInstances(projectId, location);
@@ -80,7 +80,7 @@ export let command = new Command("database:instances:list")
     return instances;
   });
 
-if (previews.rtdbmanagement) {
+if (experiments.isEnabled("rtdbmanagement")) {
   command = command.option(
     "-l, --location <location>",
     "(optional) location for the database instance, defaults to us-central1"