+++ src/commands/ext-install.ts
@@ -30,7 +30,7 @@ import { getRandomString } from "../extensions/utils";
 import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
 import { track } from "../track";
-import { previews } from "../previews";
+import * as experiments from "../experiments";
 import { Options } from "../options";
 import * as manifest from "../extensions/manifest";
 
@@ -44,7 +44,7 @@ marked.setOptions({
 export const command = new Command("ext:install [extensionName]")
   .description(
     "install an official extension if [extensionName] or [extensionName@version] is provided; " +
-      (previews.extdev
+      (experiments.isEnabled("extdev")
         ? "install a local extension if [localPathOrUrl] or [url#root] is provided; install a published extension (not authored by Firebase) if [publisherId/extensionId] is provided "
         : "") +
       "or run with `-i` to see all available extensions."